feat: react-scripts. allow parametrise webpack config path (#16644)
