+++ npm/react/plugins/react-scripts/index.js
@@ -1,9 +1,21 @@
 const { startDevServer } = require('@cypress/webpack-dev-server')
 const findReactScriptsWebpackConfig = require('./findReactScriptsWebpackConfig')
 
-module.exports = (on, config) => {
+module.exports = (
+  on,
+  config, {
+    webpackConfigPath,
+  } = {
+    webpackConfigPath: 'react-scripts/config/webpack.config',
+  },
+) => {
   on('dev-server:start', async (options) => {
-    return startDevServer({ options, webpackConfig: findReactScriptsWebpackConfig(config) })
+    return startDevServer({
+      options,
+      webpackConfig: findReactScriptsWebpackConfig(config, {
+        webpackConfigPath,
+      }),
+    })
   })
 
   config.env.reactDevtools = true
